Move TypesFactory to symbols package
